+++ mobile/android/android-components/components/support/images/src/main/java/mozilla/components/support/images/ImageRequest.kt
@@ -4,23 +4,20 @@
 
 package mozilla.components.support.images
 
-import androidx.annotation.DimenRes
+import androidx.annotation.Px
+
+/**
+ * A request to save an image. This is an alias for the id of the image.
+ */
+typealias ImageSaveRequest = String
 
 /**
  * A request to load an image.
  *
  * @property id The id of the image to retrieve.
- * @property size The preferred size of the image that should be loaded.
+ * @property size The preferred size of the image that should be loaded in pixels.
  */
-data class ImageRequest(
+data class ImageLoadRequest(
     val id: String,
-    val size: Size = Size.DEFAULT
-) {
-
-    /**
-     * Supported sizes.
-     */
-    enum class Size(@DimenRes val dimen: Int) {
-        DEFAULT(R.dimen.mozac_support_images_size_default)
-    }
-}
+    @Px val size: Int
+)